backend/requirements.txt
asgiref==3.3.4
astroid==2.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.3.1
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==2.10
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
